+++ finchge/fitness/__init__.py
@@ -2,6 +2,7 @@ from finchge.fitness.fitness_evaluator import FitnessEvaluator
 from finchge.fitness.fitness_functions import (
     AccuracyFitness,
     GEFitnessFunction,
+    MAEFitness,
     RewardFitness,
     RMSEFitness,
     StringMatchFitness,
@@ -13,5 +14,6 @@ __all__ = [
     "AccuracyFitness",
     "RewardFitness",
     "RMSEFitness",
+    "MAEFitness",
     "StringMatchFitness",
 ]