+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## [1.0.1-beta.7] - 2026-04-05
+
+### Added
+- Added API docs and documentation for benchmarks
+
+### Changed
+- Revised Introduction section of documentation to include links to API docs.
+
+### Fixed
+- Fixed grammar parser to support angular brackets in phenotype
+- Updated Cartpole grammar to make more readable with comparison symbols.
+
 ## [1.0.1-beta.6] - 2026-04-04
 
 ### Added